+++ packages/extensions/pi-peon-ping/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `findPeonCli` / `execPeonCli` utility functions for CLI invocation of peon.sh
+- `peon-ping-toggle` and `peon-ping-use` commands (registerCommand)
+
 ## [1.0.0] - 2025-07-25
 
 ### Added